start converting to markdown for indexes.
note:  there is an error in the indexing in the original document.  Under "Marshalling Siege Combat", I.C.2, the second element is listed as "1.b.", where everything else is indexed by letters at that level.  This should be "b.".  It appears that the indexing was done manually for this document.